Simplify scm/git data flow
