+++ packages/git/src/browser/git-repository-provider.spec.ts
@@ -31,6 +31,7 @@ import { GitRepositoryProvider } from './git-repository-provider';
 import * as sinon from 'sinon';
 import * as chai from 'chai';
 import { ScmService } from '@theia/scm/lib/browser';
+import { GitCommitMessageValidator } from './git-commit-message-validator';
 const expect = chai.expect;
 
 disableJSDOM();
@@ -67,6 +68,7 @@ describe('GitRepositoryProvider', () => {
     let mockStorageService: StorageService;
 
     let gitRepositoryProvider: GitRepositoryProvider;
+    let scmService: ScmService;
     const mockRootChangeEmitter: Emitter<FileStat[]> = new Emitter();
     const mockFileChangeEmitter: Emitter<FileChange[]> = new Emitter();
 
@@ -92,6 +94,7 @@ describe('GitRepositoryProvider', () => {
         testContainer.bind(FileSystemWatcher).toConstantValue(mockFileSystemWatcher);
         testContainer.bind(StorageService).toConstantValue(mockStorageService);
         testContainer.bind(ScmService).toSelf().inSingletonScope();
+        testContainer.bind(GitCommitMessageValidator).toSelf().inSingletonScope();
 
         sinon.stub(mockWorkspaceService, 'onWorkspaceChanged').value(mockRootChangeEmitter.event);
         sinon.stub(mockFileSystemWatcher, 'onFilesChanged').value(mockFileChangeEmitter.event);
@@ -105,6 +108,7 @@ describe('GitRepositoryProvider', () => {
         sinon.stub(mockWorkspaceService, 'roots').value(Promise.resolve());
         (<sinon.SinonStub>mockWorkspaceService.tryGetRoots).returns(roots);
         gitRepositoryProvider = testContainer.get<GitRepositoryProvider>(GitRepositoryProvider);
+        scmService = testContainer.get<ScmService>(ScmService);
         (<sinon.SinonStub>mockFilesystem.exists).resolves(true);
         (<sinon.SinonStub>mockGit.repositories).withArgs(folderA.uri, {}).resolves(allRepos);
 
@@ -128,19 +132,20 @@ describe('GitRepositoryProvider', () => {
         stubWsRoots.onCall(1).returns(oldRoots);
         stubWsRoots.onCall(2).returns(newRoots);
         gitRepositoryProvider = testContainer.get<GitRepositoryProvider>(GitRepositoryProvider);
+        scmService = testContainer.get<ScmService>(ScmService);
         (<sinon.SinonStub>mockFilesystem.exists).resolves(true);
         (<sinon.SinonStub>mockGit.repositories).withArgs(folderA.uri, {}).resolves(allReposA);
         (<sinon.SinonStub>mockGit.repositories).withArgs(folderB.uri, {}).resolves(allReposB);
 
         let counter = 0;
-        gitRepositoryProvider.onDidChangeRepository(selected => {
+        scmService.onDidAddRepository(selected => {
             counter++;
             if (counter === 3) {
                 expect(gitRepositoryProvider.allRepositories.length).to.eq(allReposA.concat(allReposB).length);
                 expect(gitRepositoryProvider.allRepositories[0].localUri).to.eq(allReposA[0].localUri);
                 expect(gitRepositoryProvider.allRepositories[1].localUri).to.eq(allReposA[1].localUri);
                 expect(gitRepositoryProvider.allRepositories[2].localUri).to.eq(allReposB[0].localUri);
-                expect(selected && selected.localUri).to.eq(allReposA[0].localUri);
+                expect(selected && selected.provider.rootUri).to.eq(allReposB[0].localUri);
                 done();
             }
         });
@@ -164,19 +169,20 @@ describe('GitRepositoryProvider', () => {
         stubWsRoots.onCall(1).returns(oldRoots);
         stubWsRoots.onCall(2).returns(newRoots);
         gitRepositoryProvider = testContainer.get<GitRepositoryProvider>(GitRepositoryProvider);
+        scmService = testContainer.get<ScmService>(ScmService);
         (<sinon.SinonStub>mockFilesystem.exists).resolves(true);
         (<sinon.SinonStub>mockGit.repositories).withArgs(folderA.uri, {}).resolves(allReposA);
         (<sinon.SinonStub>mockGit.repositories).withArgs(folderB.uri, {}).resolves(allReposB);
 
         let counter = 0;
-        gitRepositoryProvider.onDidChangeRepository(selected => {
+        scmService.onDidAddRepository(selected => {
             counter++;
             if (counter === 3) {
                 expect(gitRepositoryProvider.allRepositories.length).to.eq(allReposA.concat(allReposB).length);
                 expect(gitRepositoryProvider.allRepositories[0].localUri).to.eq(allReposA[0].localUri);
                 expect(gitRepositoryProvider.allRepositories[1].localUri).to.eq(allReposA[1].localUri);
                 expect(gitRepositoryProvider.allRepositories[2].localUri).to.eq(allReposB[0].localUri);
-                expect(selected && selected.localUri).to.eq(allReposA[0].localUri);
+                expect(selected && selected.provider.rootUri).to.eq(allReposB[0].localUri);
                 done();
             }
         });
