+++ packages/git/src/browser/git-repository-provider.ts
@@ -14,16 +14,24 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-import { Git, Repository } from '../common';
+import { CommitWithChanges, Git, Repository } from '../common';
 import { injectable, inject } from 'inversify';
 import { WorkspaceService } from '@theia/workspace/lib/browser/workspace-service';
 import { FileSystem, FileStat } from '@theia/filesystem/lib/common';
-import { DisposableCollection, Event, Emitter } from '@theia/core';
-import { StorageService } from '@theia/core/lib/browser';
+import { DisposableCollection, Emitter, Event } from '@theia/core';
 import URI from '@theia/core/lib/common/uri';
 import { FileSystemWatcher } from '@theia/filesystem/lib/browser/filesystem-watcher';
-import { ScmService } from '@theia/scm/lib/browser';
+import {
+    ScmAmendSupport,
+    ScmCommand,
+    ScmCommit,
+    ScmProvider,
+    ScmRepository,
+    ScmResourceGroup,
+    ScmService
+} from '@theia/scm/lib/browser';
 import debounce = require('lodash.debounce');
+import { GitCommitMessageValidator } from './git-commit-message-validator';
 
 export interface GitRefreshOptions {
     readonly maxCount: number
@@ -32,11 +40,7 @@ export interface GitRefreshOptions {
 @injectable()
 export class GitRepositoryProvider {
 
-    protected _selectedRepository: Repository | undefined;
-    protected _allRepositories?: Repository[];
     protected readonly onDidChangeRepositoryEmitter = new Emitter<Repository | undefined>();
-    protected readonly selectedRepoStorageKey = 'theia-git-selected-repository';
-    protected readonly allRepoStorageKey = 'theia-git-all-repositories';
 
     constructor(
         @inject(Git) protected readonly git: Git,
@@ -44,13 +48,19 @@ export class GitRepositoryProvider {
         @inject(FileSystemWatcher) protected readonly watcher: FileSystemWatcher,
         @inject(FileSystem) protected readonly fileSystem: FileSystem,
         @inject(ScmService) protected readonly scmService: ScmService,
-        @inject(StorageService) protected readonly storageService: StorageService
+        @inject(GitCommitMessageValidator) protected readonly commitMessageValidator: GitCommitMessageValidator
     ) {
         this.initialize();
     }
 
     protected readonly toDisposeOnWorkspaceChange = new DisposableCollection();
     protected async initialize(): Promise<void> {
+        this.scmService.onDidChangeSelectedRepositories(scmRepository => {
+            if (scmRepository && scmRepository.provider.contextValue === 'Git') {
+                this.onDidChangeRepositoryEmitter.fire({ localUri: scmRepository.provider.rootUri });
+            }
+        });
+
         /*
          * Listen for changes to the list of workspaces.  This will not be fired when changes
          * are made inside a workspace.
@@ -71,9 +81,7 @@ export class GitRepositoryProvider {
             this.lazyRefresh();
         });
 
-        this._selectedRepository = await this.storageService.getData<Repository | undefined>(this.selectedRepoStorageKey);
-        this._allRepositories = await this.storageService.getData<Repository[]>(this.allRepoStorageKey);
-        if (!this._allRepositories) {
+        if (!(this.scmService.repositories.length === 0)) {
             await this.refresh({ maxCount: 1 });
         }
         await this.refresh();
@@ -87,37 +95,35 @@ export class GitRepositoryProvider {
      * If no repositories are available, returns `undefined`.
      */
     get selectedRepository(): Repository | undefined {
-        return this._selectedRepository;
+        if (this.scmService.selectedRepository) {
+            return { localUri: this.scmService.selectedRepository.provider.rootUri };
+        }
     }
 
     /**
      * Sets or un-sets the repository.
      */
     set selectedRepository(repository: Repository | undefined) {
-        if (!repository) {
+        if (repository) {
+            this.scmService.selectedRepository = this.scmService.repositories.find(scmRepository => scmRepository.provider.rootUri === repository.localUri);
+        } else {
             this.scmService.selectedRepository = undefined;
         }
-        this._selectedRepository = repository;
-        this.storageService.setData<Repository | undefined>(this.selectedRepoStorageKey, repository);
-        this.fireDidChangeRepository();
     }
 
     get onDidChangeRepository(): Event<Repository | undefined> {
         return this.onDidChangeRepositoryEmitter.event;
-    }
-    protected fireDidChangeRepository(): void {
-        this.onDidChangeRepositoryEmitter.fire(this._selectedRepository);
     }
 
     /**
      * Returns with all know repositories.
      */
     get allRepositories(): Repository[] {
-        return this._allRepositories || [];
+        return this.scmService.repositories.map(scmRepository => ({ localUri: scmRepository.provider.rootUri }));
     }
 
     findRepository(uri: URI): Repository | undefined {
-        const reposSorted = this._allRepositories ? this._allRepositories.sort(Repository.sortComparator) : [];
+        const reposSorted = this.scmService.repositories.map(scmRepo => ({ localUri: scmRepo.provider.rootUri })).sort(Repository.sortComparator);
         return reposSorted.find(repo => new URI(repo.localUri).isEqualOrParent(uri));
     }
 
@@ -147,27 +153,197 @@ export class GitRepositoryProvider {
                 roots.push(root);
             }
         }
-        const repoUris = new Map<string, Repository>();
+        const repoUris: string[] = [];
         const reposOfRoots = await Promise.all(
             roots.map(r => this.git.repositories(r.uri, { ...options }))
         );
         reposOfRoots.forEach(reposPerRoot => {
             reposPerRoot.forEach(repoOfOneRoot => {
-                repoUris.set(repoOfOneRoot.localUri, repoOfOneRoot);
+                repoUris.push(repoOfOneRoot.localUri);
             });
         });
-        this._allRepositories = Array.from(repoUris.values());
-        this.storageService.setData<Repository[]>(this.allRepoStorageKey, this._allRepositories);
-        const selectedRepository = this._selectedRepository;
-        if (!selectedRepository || !this.exists(selectedRepository)) {
-            this.selectedRepository = this._allRepositories[0];
-        } else {
-            this.fireDidChangeRepository();
+
+        repoUris.forEach(uri => {
+            if (!this.scmService.repositories.find(scmRepository => scmRepository.provider.rootUri === uri)) {
+                this.registerScmProvider(uri);
+            }
+        });
+        this.scmService.repositories.forEach(scmRepository => {
+            if (repoUris.indexOf(scmRepository.provider.rootUri) < 0) {
+                scmRepository.dispose();
+            }
+        });
+    }
+
+    protected registerScmProvider(uri: string): ScmRepository | undefined {
+        const amendSupport: ScmAmendSupport = new GitAmendSupport({ localUri: uri }, this.git);
+        const provider = new ScmProviderImpl('Git', uri.substring(uri.lastIndexOf('/') + 1), uri, amendSupport);
+        const repo = this.scmService.registerScmProvider(provider);
+        repo.input.placeholder = 'Commit message';
+        repo.input.validateInput = async input => {
+            const validate = await this.commitMessageValidator.validate(input);
+            if (validate) {
+                const { message, status } = validate;
+                return { message, type: status };
+            }
+        };
+        return repo;
+    }
+
+}
+
+export class GitAmendSupport implements ScmAmendSupport {
+
+    constructor(protected readonly repository: Repository, protected readonly git: Git) { }
+
+    public async getInitialAmendingCommits(amendingHeadCommitSha: string, latestCommitSha: string): Promise<ScmCommit[]> {
+        const commits = await this.git.log(
+            this.repository,
+            {
+                range: { toRevision: amendingHeadCommitSha, fromRevision: latestCommitSha },
+                maxCount: 50
+            }
+        );
+
+        return commits.map(this.createScmCommit);
+    }
+
+    public async getMessage(commit: string): Promise<string> {
+        return (await this.git.exec(this.repository, ['log', '-n', '1', '--format=%B', commit])).stdout.trim();
+    }
+
+    public async reset(commit: string): Promise<void> {
+        await this.git.exec(this.repository, ['reset', commit, '--soft']);
+    }
+
+    public async getLastCommit(): Promise<ScmCommit | undefined> {
+        const commits = await this.git.log(this.repository, { maxCount: 1 });
+        if (commits.length > 0) {
+            return this.createScmCommit(commits[0]);
         }
     }
 
-    protected exists(repository: Repository): boolean {
-        return !!this._allRepositories && this._allRepositories.some(repository2 => Repository.equal(repository, repository2));
+    private createScmCommit(gitCommit: CommitWithChanges) {
+        return {
+            id: gitCommit.sha,
+            summary: gitCommit.summary,
+            authorName: gitCommit.author.name,
+            authorEmail: gitCommit.author.email,
+            authorDateRelative: gitCommit.authorDateRelative
+        };
+    }
+}
+
+export class ScmProviderImpl implements ScmProvider {
+    private static ID = 0;
+
+    private onDidChangeEmitter = new Emitter<void>();
+    private onDidChangeResourcesEmitter = new Emitter<void>();
+    private onDidChangeCommitTemplateEmitter = new Emitter<string>();
+    private onDidChangeStatusBarCommandsEmitter = new Emitter<ScmCommand[]>();
+    private disposableCollection: DisposableCollection = new DisposableCollection();
+    private _groups: ScmResourceGroup[];
+    private _count: number | undefined;
+    readonly handle = 0;
+
+    constructor(
+        private _contextValue: string,
+        private _label: string,
+        private _rootUri: string,
+        private _amendSupport: ScmAmendSupport,
+    ) {
+        this.disposableCollection.push(this.onDidChangeEmitter);
+        this.disposableCollection.push(this.onDidChangeResourcesEmitter);
+        this.disposableCollection.push(this.onDidChangeCommitTemplateEmitter);
+        this.disposableCollection.push(this.onDidChangeStatusBarCommandsEmitter);
     }
 
+    private _id = `scm${ScmProviderImpl.ID++}`;
+
+    get id(): string {
+        return this._id;
+    }
+    get groups(): ScmResourceGroup[] {
+        return this._groups;
+    }
+
+    set groups(groups: ScmResourceGroup[]) {
+        this._groups = groups;
+    }
+
+    get label(): string {
+        return this._label;
+    }
+
+    get rootUri(): string {
+        return this._rootUri;
+    }
+
+    get contextValue(): string {
+        return this._contextValue;
+    }
+
+    get onDidChangeResources(): Event<void> {
+        return this.onDidChangeResourcesEmitter.event;
+    }
+
+    get commitTemplate(): string | undefined {
+        return undefined;
+    }
+
+    get acceptInputCommand(): ScmCommand | undefined {
+        return {
+            id: 'git.commit.all',
+            tooltip: 'Commit all the staged changes',
+            text: 'Commit',
+        };
+    }
+
+    get statusBarCommands(): ScmCommand[] | undefined {
+        return undefined;
+    }
+
+    get count(): number | undefined {
+        return this._count;
+    }
+
+    set count(count: number | undefined) {
+        this._count = count;
+    }
+
+    get onDidChangeCommitTemplate(): Event<string> {
+        return this.onDidChangeCommitTemplateEmitter.event;
+    }
+
+    get onDidChangeStatusBarCommands(): Event<ScmCommand[]> {
+        return this.onDidChangeStatusBarCommandsEmitter.event;
+    }
+
+    get onDidChange(): Event<void> {
+        return this.onDidChangeEmitter.event;
+    }
+
+    dispose(): void {
+        this.disposableCollection.dispose();
+    }
+
+    async getOriginalResource(uri: URI): Promise<URI | undefined> {
+        return undefined;
+    }
+
+    fireChangeStatusBarCommands(commands: ScmCommand[]): void {
+        this.onDidChangeStatusBarCommandsEmitter.fire(commands);
+    }
+
+    fireChangeResources(): void {
+        this.onDidChangeResourcesEmitter.fire(undefined);
+    }
+
+    fireChange(): void {
+        this.onDidChangeEmitter.fire(undefined);
+    }
+
+    get amendSupport(): ScmAmendSupport | undefined {
+        return this._amendSupport;
+    }
 }
