+++ packages/git/src/browser/git-repository-tracker.ts
@@ -20,6 +20,7 @@ import { Event, Emitter, DisposableCollection } from '@theia/core';
 import { GitRepositoryProvider } from './git-repository-provider';
 import { GitWatcher, GitStatusChangeEvent } from '../common/git-watcher';
 import URI from '@theia/core/lib/common/uri';
+import { ScmService } from '@theia/scm/lib/browser';
 
 /**
  * The repository tracker watches the selected repository for status changes. It provides a convenient way to listen on status updates.
@@ -35,23 +36,25 @@ export class GitRepositoryTracker {
         @inject(Git) protected readonly git: Git,
         @inject(GitRepositoryProvider) protected readonly repositoryProvider: GitRepositoryProvider,
         @inject(GitWatcher) protected readonly gitWatcher: GitWatcher,
+        @inject(ScmService) protected readonly scmService: ScmService
     ) { }
 
     @postConstruct()
     protected async init() {
-        this.repositoryProvider.onDidChangeRepository(async repository => {
+        this.scmService.onDidChangeSelectedRepositories(async repository => {
             this.workingDirectoryStatus = undefined;
             this.toDispose.dispose();
-            if (repository) {
-                this.toDispose.push(await this.gitWatcher.watchGitChanges(repository));
+            if (repository && repository.provider.contextValue === 'Git') {
+                const gitRepository = { localUri: repository.provider.rootUri };
+                this.toDispose.push(await this.gitWatcher.watchGitChanges(gitRepository));
                 this.toDispose.push(this.gitWatcher.onGitEvent((event: GitStatusChangeEvent) => {
                     this.workingDirectoryStatus = event.status;
                     this.onGitEventEmitter.fire(event);
                 }));
-                this.workingDirectoryStatus = await this.git.status(repository);
+                this.workingDirectoryStatus = await this.git.status(gitRepository);
             }
         });
-        if (this.repositoryProvider.allRepositories.length === 0) {
+        if (this.scmService.repositories.length === 0) {
             await this.repositoryProvider.refresh();
         }
         if (this.selectedRepository) {
@@ -78,13 +81,6 @@ export class GitRepositoryTracker {
      */
     get selectedRepositoryStatus(): WorkingDirectoryStatus | undefined {
         return this.workingDirectoryStatus;
-    }
-
-    /**
-     * Emits when the selected repository has changed.
-     */
-    get onDidChangeRepository(): Event<Repository | undefined> {
-        return this.repositoryProvider.onDidChangeRepository;
     }
 
     /**
